fix fgbg

+++ srcs/builtin/jobs.c
@@ -6,45 +6,24 @@
 /*   By: aardjoun <aardjoun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/02/28 15:18:15 by aardjoun          #+#    #+#             */
-/*   Updated: 2014/03/18 17:45:26 by lsolofri         ###   ########.fr       */
+/*   Updated: 2014/03/21 12:01:05 by aardjoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/msh.h"
-/*
-void		new_job(char *name, int job, int pid, int status)
-{
-	t_jobs	*jobs;
 
-	jobs->name = ft_strdup(name);
-	jobs->job = job;
-	job->pid = pid;
-	job->status = status;
-	job->next = NULL;
-}
-
-t_jobs		*add_job(t_jobs *jobs, char *name, int job, int pid, int status)
-{
-	t_jobs	*tmp;
-
-	tmp = new_job(name, job, pid, status);
-	if (jobs == NULL)
-		jobs = tmp;
-	else
-		while (jobs->next != NULL)
-			jobs = jobs->next;
-	tmp->next = jobs;
-	return (jobs);
-}
-*/
 void		fg_bg(t_jobs *jobs, char ** tab, int k)
 {
 	int		rt;
 	int		i;
 
-	(void)k;
 	rt = 0;
 	i = 0;
+	if (jobs == NULL)
+	{
+		job_error(k);
+		return ;
+	}
 	if (tab[1])
 	{
 		while (tab[1] && tab[1][++i] != '\0')
@@ -61,7 +40,9 @@ void		fg_bg(t_jobs *jobs, char ** tab, int k)
 		while (jobs->next != NULL)
 			jobs = jobs->next;
 		kill(jobs->pid, SIGCONT);
-		wait(&jobs->pid);
+		g_jobs = remove_jobs(g_jobs, jobs->pid);
+		if (k == 1)
+			wait(&jobs->pid);
 	}
 }
 
@@ -70,9 +51,17 @@ int			search_job(t_jobs *jobs, char *tab, int rt)
 	(void)tab;
 	(void)rt;
 	while (jobs->next != NULL)
-		jobs = jobs->next;
-	kill(jobs->pid, SIGCONT);
-	wait(&jobs->pid);
-	g_jobs = remove_jobs(g_jobs, jobs->pid);
+	{
+		if (jobs->pid != ft_atoi(tab)) 
+			jobs = jobs->next;
+	}
+	if (jobs->pid == ft_atoi(tab))
+	{
+		kill(jobs->pid, SIGCONT);
+		wait(&jobs->pid);
+		g_jobs = remove_jobs(g_jobs, jobs->pid);
+	}
+	else
+		pid_error(tab);
 	return (0);
 }
